+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - BREAKING: place the 2 optional args to cursor helpers in an options object
 - add support for simple managed migrations
 - add signal promises for opened, closed, outdated and blocked
+- add errorevent param to cursor error handler
 - add doc comments to all APIs for in-editor help
 
 ## 2.0.0 - 2020-08-21